chore: collapse schema V1/V2 into V3 — remove migration infrastructure
Move all model definitions (ChatMessage, ChatSession, SamplerPreset,
APIEndpoint) directly into BaseChatSchemaV3, eliminating the V1 and V2
versioned schemas and the BaseChatMigrationPlan migration stages.

Safe to do now while the repo is private with no users carrying V1/V2
data. The schema versioning system remains in place for future versions.
