+++ Sources/BaseChatCore/ModelContainerFactory.swift
@@ -1,9 +1,8 @@
 import SwiftData
 
-/// Creates `ModelContainer` instances pre-configured with ``BaseChatMigrationPlan``.
+/// Creates `ModelContainer` instances configured with the current schema.
 ///
-/// Use `ModelContainerFactory` instead of constructing `ModelContainer` by hand
-/// so that your store is automatically migrated whenever the schema changes.
+/// Use `ModelContainerFactory` instead of constructing `ModelContainer` by hand.
 ///
 /// ```swift
 /// // On-disk store (typical app setup)
@@ -13,29 +12,28 @@ import SwiftData
 /// let container = try ModelContainerFactory.makeInMemoryContainer()
 /// ```
 public enum ModelContainerFactory {
-    /// The newest schema version that should be opened for application stores.
+    /// The current schema version.
     public static var currentSchema: any VersionedSchema.Type {
-        BaseChatMigrationPlan.schemas.last ?? BaseChatSchemaV1.self
+        BaseChatSchemaV3.self
     }
 
-    /// Returns an on-disk `ModelContainer` configured with ``BaseChatMigrationPlan``.
+    /// Returns an on-disk `ModelContainer` configured with the current schema.
     ///
     /// - Parameter configurations: Additional `ModelConfiguration` values to
     ///   pass to `ModelContainer`. Defaults to a single default (on-disk) config.
-    /// - Returns: A `ModelContainer` whose store will be automatically migrated.
+    /// - Returns: A `ModelContainer` using the current schema.
     /// - Throws: If `ModelContainer` initialisation fails.
     public static func makeContainer(
         configurations: [ModelConfiguration] = [ModelConfiguration()]
     ) throws -> ModelContainer {
         try ModelContainer(
-            for: Schema(versionedSchema: currentSchema),
-            migrationPlan: BaseChatMigrationPlan.self,
+            for: Schema(versionedSchema: BaseChatSchemaV3.self),
             configurations: configurations
         )
     }
 
-    /// Returns an ephemeral in-memory `ModelContainer` configured with
-    /// ``BaseChatMigrationPlan``.
+    /// Returns an ephemeral in-memory `ModelContainer` configured with the
+    /// current schema.
     ///
     /// Suitable for tests, SwiftUI previews, and any context where data must
     /// not be persisted to disk.
