+++ Sources/BaseChatCore/Models/ChatSession.swift
@@ -1,2 +1,2 @@
-/// Public alias for the V1 SwiftData chat session model.
-public typealias ChatSession = BaseChatSchemaV1.ChatSession
+/// Public alias for the current SwiftData chat session model.
+public typealias ChatSession = BaseChatSchemaV3.ChatSession
