+++ Sources/BaseChatCore/Models/SamplerPreset.swift
@@ -1,2 +1,2 @@
-/// Public alias for the V1 SwiftData sampler preset model.
-public typealias SamplerPreset = BaseChatSchemaV1.SamplerPreset
+/// Public alias for the current SwiftData sampler preset model.
+public typealias SamplerPreset = BaseChatSchemaV3.SamplerPreset
